[PCFG-9] #done details of a CE

+++ src/main/java/it/pagopa/pagopa/apiconfig/model/Station.java
@@ -8,7 +8,6 @@ import lombok.*;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-import java.math.BigDecimal;
 
 /**
  * Station
@@ -35,7 +34,7 @@ public class Station {
     @JsonProperty("version")
     @Schema(required = true, description = "number version")
     @NotNull
-    private BigDecimal version;
+    private Long version;
 
 
 }
